+++ templates/python.flask/{{cookiecutter.project_slug}}/config/default.py
@@ -1,4 +1,7 @@
 HTTP_HOST = ''
 HTTP_PORT = 8080
 
-LOG="/var/{{cookiecutter.project_slug}}/error.log"
+LOG = "/var/{{cookiecutter.project_slug}}"
+
+# WSGI Settings
+WSGI_LOG = 'default'