+++ templates/python.flask/{{cookiecutter.project_slug}}/config/development.py
@@ -1,1 +1,4 @@
 # local development environment setting
+
+# WSGI Settings
+WSGI_LOG = 'default'