+++ templates/python.flask/{{cookiecutter.project_slug}}/config/gqc.py
@@ -1,1 +1,4 @@
 # gqc environment setting
+
+# WSGI Settings
+WSGI_LOG = None