+++ templates/python.flask/{{cookiecutter.project_slug}}/config/prd.py
@@ -1,1 +1,4 @@
 # prd environment setting
+
+# WSGI Settings
+WSGI_LOG = None