+++ templates/python.flask/{{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}/views/__init__.py
@@ -1,3 +1,4 @@
 from faq import version
+from index import index
 
-__all__ = ['version']
+__all__ = ['version', 'index']