+++ templates/python.flask/{{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}/views/index.py
@@ -1,0 +1,6 @@
+from {{cookiecutter.project_slug}} import app
+
+
+@app.route("/", methods=['GET'])
+def index():
+    return "hello, world"